Decoupled module loader from module searcher
